test(QueryBuilder): 누락된 테스트 추가
